CrewAI/requirements.txt
langchain
openai
streamlit
crewai
crewai_tools
python-dotenv